// types/schema.d.ts
export interface components {
    schemas: {
// ...
        LoanResponse: {
            /** Loan Amount */
            loan_amount: string;
            /** Interest Rate */
            interest_rate: string;
            /** Duration Months */
            duration_months: number;
            /** Id */
            id: number;
            /** User Id */
            user_id: number;
            /** Monthly Repayment */
            monthly_repayment: string;
            /** Total Repayable */
            total_repayable: string;
            /** Amount Paid */
            amount_paid: string;
            /** Balance */
            balance: string;
            status: components["schemas"]["LoanStatus"];
            /**
             * Application Date
             * Format: date-time
             */
            application_date: string;
            /** Approval Date */
            approval_date: string | null;
            /** Disbursement Date */
            disbursement_date: string | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
        UserResponse: {
            /** Member Id */
            member_id: string;
            /**
             * Email
             * Format: email
             */
            email: string;
            /** Full Name */
            full_name: string;
            /** Phone */
            phone?: string | null;
            /** @default member */
            role: components["schemas"]["UserRole"];
            /** Id */
            id: number;
            status: components["schemas"]["UserStatus"];
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}